+++ source/Manager.py
@@ -1,8 +1,8 @@
-# それぞれの金額
-goods_name_list = ["1.ワッフル", "2.クレープ", "3.豚タン", "4.ベビーカステラ"]
-goods_list = [180, 320, 500, 200]
+# # それぞれの金額
+# goods_name_list = ["1.ワッフル", "2.クレープ", "3.豚タン", "4.ベビーカステラ"]
+# goods_list = [180, 320, 500, 200]
 
-def sum_checkout(goods_count_list):
+def sum_checkout(goods_name_list, goods_list, goods_count_list):
     print(goods_count_list)
     all_sum_amount = 0
     for x in range(4):
@@ -10,25 +10,25 @@ def sum_checkout(goods_count_list):
     return all_sum_amount
 
 
-def manager_num(manager_code):
+def manager_num(manager_code, goods_name_list, goods_list, goods_count_list):
     if manager_code == 1:
-        code1_event()
-        print_goods()
+        code1_event(goods_name_list, goods_list, goods_count_list)
+        print_goods(goods_name_list, goods_list, goods_count_list)
     elif manager_code == 2:
-        code2_event()
-        print_goods()
+        code2_event(goods_name_list, goods_list, goods_count_list)
+        print_goods(goods_name_list, goods_list, goods_count_list)
     elif manager_code == 3:
-        code3_event()
+        code3_event(goods_name_list, goods_list, goods_count_list)
     else:
         print("管理者コードが正しくありません。")
         
-def code1_event(goods_count_list):
+def code1_event(goods_name_list, goods_list, goods_count_list):
     print("===")
     for x in range(4):
         goods_count_list[x] = 0
     print("売上をリセットしました。\n")
     
-def code2_event():
+def code2_event(goods_name_list, goods_list, goods_count_list):
     print("===")
     change_number = int(input("価格を変更する商品の番号を入力してください。 >"))
     if change_number <= 0 or change_number > 4:
@@ -41,14 +41,14 @@ def code2_event():
             goods_list[change_number - 1] = change_amount
             print("変更しました。\n")
             
-def code3_event(goods_count_list):
+def code3_event(goods_name_list, goods_list, goods_count_list):
     get_enter = input("Enterキー押下でタイトル画面の表示に戻る\n")
     if get_enter == "":
         # ここでメインに戻るーーーーー
         from Ticket_machine import main
-        main(goods_count_list)
+        main(goods_name_list, goods_list, goods_count_list)
 
-def print_goods(goods_count_list):
+def print_goods(goods_name_list, goods_list, goods_count_list):
     print("======= 商品一覧 =======")
     print("商品      単価  販売数  売上金額")
     print("=======================")
@@ -57,7 +57,7 @@ def print_goods(goods_count_list):
         print(f"{goods_name_list[num]} {goods_list[num]}円  {goods_count_list[num]}  {goods_list[num]*goods_count_list[num]}")
         
     print("———")
-    print(f"総売上金額  {sum_checkout(goods_count_list)}円\n")
+    print(f"総売上金額  {sum_checkout(goods_name_list, goods_list, goods_count_list)}円\n")
 
     print("=== 管理メニュー ====")
     print("1. 売上をリセットする")
@@ -68,15 +68,14 @@ def print_goods(goods_count_list):
     manager_code = int(input("管理コード入力:"))
 
     # 不正な受け取りに対するエラー文も出せるようにしておこう！！
-    manager_num(manager_code)
+    manager_num(manager_code, goods_name_list, goods_list, goods_count_list)
     
 # 表示
-def manager(goods_count_list):
-    print(goods_count_list)
+def manager(goods_name_list, goods_list, goods_count_list):
     print ("***********************")
     print("       管理画面        ")
     print ("***********************\n")
-    print_goods(goods_count_list)
+    print_goods(goods_name_list, goods_list, goods_count_list)
